Final project Add MiddleWares-errorHandling,Ratting Change names of functions to __Async Correct function to placeOrder Add validationds for UserDto object Change from ₪ to $

+++ DTO/OrderDto.cs
@@ -12,7 +12,7 @@ namespace DTO
 
         public DateTime OrderDate { get; set; }
 
-        public decimal OrderSum { get; set; }
+        public double OrderSum { get; set; }
 
         public int UserId { get; set; }
